feat: added Command STFU
