+++ Rolemanager/RoleCommands.cs
@@ -154,43 +154,5 @@ namespace DiscordBot.Rolemanager
       }
       return false;
     }
-
-    /*
-      public class MuteCommand : BaseCommandModule
-  {
-    Prüfung der Berechtigung und Cooldown für den Befehl.
-    [Command("stfu"), RequireUserPermissions(Permissions.Administrator) /*Cooldown(1, 30, CooldownBucketType.Channel)]
-
-    public async Task STFU(CommandContext ctx)
-    {
-      MuteCommandTable mcTable = new MuteCommandTable();
-      Was passieren sollte: Alle User in dem Sprachkanal des ausführenden Users, sollen gemuted werden.
-                            Ausgenommen davon ist der ausführende User.
-                            Nach ablauf eines weiteren Cooldowns werden die User wieder entmuted.
-      await ctx.RespondAsync("Silence!");
-      mcTable.arrayChannels = new List<DiscordChannel>(ctx.Guild.Channels.Values).ToArray();
-      foreach (var arrayChannel in mcTable.arrayChannels)
-      {
-        mcTable.arrayMembers = new List<DiscordMember>(arrayChannel.Users).ToArray();
-        foreach (var arrayMember in mcTable.arrayMembers)
-        {
-          if (arrayMember.Id != ctx.Member.Id && arrayMember.IsMuted == false)
-          {
-            await arrayMember.SetMuteAsync(true);
-            await ctx.RespondAsync("Silence!");
-          }
-          else if (arrayMember.IsMuted)
-          {
-            await arrayMember.SetMuteAsync(false);
-            await ctx.RespondAsync("You now can speak again!!111");
-          }
-          else
-          {
-            await arrayMember.SetMuteAsync(false);
-          }
-        }
-      }
-    }
-     */
   }
 }
